Fix query NOT bug to return results mirroring Mongo
Previously NOT processing was handled by applying NOT to each
comparison in an atomic manner.  This created a difference in the way
the SQL engine and the post hoc matcher were handling documents that
contained arrays.

Through more thorough testing with MongoDB it was also uncovered that
NOT processing should specifically be handled as the NOT of the
operator rather than the opposite operator OR-ed with NULL.  So the
result set for a query like { pet : { $not : { $eq : "cat" } } } will
return all of documents that do NOT exist in the result set of
documents that are returned for the query { pet : { $eq : "cat" } }.

Based on this finding along with the "documents with arrays processing"
issue, the logical conclusion was to implement NOT processing in the
SQL engine by using sub-SELECT statements and a NOT IN...  So the query
{ pet : { $not : { $eq : "cat" } } } would translate to a WHERE clause
looking like _id NOT IN (SELECT _id FROM index_table_name WHERE pet =
"cat"). The post hoc matcher code was then changed to handle array
processing w.r.t. NOT with the array as the single unit rather than
each element in the array as the single unit.  So the array comparison
was NOT-ted once the entire array in the document was processed rather
than NOT-ting each element in the array which is what caused our
initial problem.

These changes brought the matcher and the SQL engine results into synch
as well as made our Query results match the results from MongoDB.

Change details:

- Fix logic in the SQL translator to handle NOT processing correctly.
- Fix logic in the post hoc matcher to handle NOT processing correctly.
- Update tests results to reflect results as they would exist in
MongoDB.
- Add additional tests to more thoroughly test the NOT logic.
Specifically as it relates to arrays in documents.
- Remove two unnecessary checks for sqlOperator being null.  This
condition will be captured during query normalization therefore an
invalid operator should not be encountered in the translator code.
- Remove two tests that test for unsupported SQL operators and replaced
them with two validation tests that test for invalid query operators.
- Update readme.md to fix old NOT logic documentation.
- Update changelog.md to detail new NOT processing and to include
examples.