Add support for $nin operator
- Update Validator to include $nin support
- Add Validator tests for $nin.  There is no need to add any SQL
Translator tests or any Executor tests or any Matcher tests since
$nin normalizes to $not..$eq and there are tests already in place for
that.
- Move shorthand processing before NOT processing in normalization.
Thus making shorthand processing an individual unit of work in
the normalization process instead of a sub-unit of NOT processing as it
was previously.
- Remove redundant predicate validation throughout Translator since
predicate validation is already handled as part of the normalization
process.
- Simplify the placeholder generation logic during WHERE clause
construction.
- Remove redundant Float validation throughout the Matcher since this
is already handled as part of the normalization process.